unpack and convert
